Add new BannerCloseButton component
