+++ src/AppBanner/index.tsx
@@ -1,36 +1,18 @@
 import React, { useState, ReactElement } from 'react';
 import { css } from '@emotion/react';
 import { ThemeProvider } from '@emotion/react';
-import {
-    Button,
-    buttonThemeReaderRevenueBrandAlt,
-    SvgCross,
-} from '@guardian/source-react-components';
+import { Button, buttonThemeReaderRevenueBrandAlt } from '@guardian/source-react-components';
+import { space, from, until } from '@guardian/source-foundations';
 import type { TrackClick } from '../utils/tracking';
 
 import { AppStore } from '../assets/app-store';
 import { PlayStore } from '../assets/play-store';
-import { BannerColorStyles } from '../styles/colorData';
+import { BannerColorStyles, CloseButtonColorStyles } from '../styles/colorData';
 import { selfServeStyles } from '../styles/bannerCommon';
-import { useEscapeShortcut, OnCloseClick, CLOSE_BUTTON_ID } from '../bannerCommon/bannerActions';
-import { styles } from './styles';
+import { BannerCloseButton, OnCloseClick } from '../components/BannerCloseButton';
 
 import { canRender, COMPONENT_NAME } from './canRender';
 export { COMPONENT_NAME };
-
-const defaultColors: BannerColorStyles = {
-    styleBackground: '#ebe8e8',
-    styleHeader: `#333333`,
-    styleBody: '#666',
-    styleHighlight: `#333333`,
-    styleHighlightBackground: '#ebe8e8',
-    styleButton: '#ffffff',
-    styleButtonBackground: '#052962',
-    styleButtonHover: '#234b8a',
-    styleClose: `#333333`,
-    styleCloseBackground: '#ebe8e8',
-    styleCloseHover: '#ffd213',
-};
 
 export type BrazeMessageProps = {
     ophanComponentId?: string;
@@ -57,25 +39,9 @@ export const AppBanner = (props: Props): ReactElement | null => {
         trackClick,
     } = props;
 
-    const commonStyles = selfServeStyles(props.brazeMessageProps, defaultColors);
+    const styles = selfServeStyles(props.brazeMessageProps, defaultColors);
 
     const [showBanner, setShowBanner] = useState(true);
-
-    const onCloseClick: OnCloseClick = (evt, internalButtonId) => {
-        evt.preventDefault();
-        onCloseAction(internalButtonId);
-    };
-
-    const onCloseAction = (internalButtonId: number): void => {
-        setShowBanner(false);
-        document.body.focus();
-        trackClick({
-            internalButtonId,
-            ophanComponentId: ophanComponentId as string,
-        });
-    };
-
-    useEscapeShortcut(() => onCloseAction(CLOSE_BUTTON_ID));
 
     // This is to keep button colors the same as before
     // https://github.com/guardian/braze-components/pull/123
@@ -98,27 +64,38 @@ export const AppBanner = (props: Props): ReactElement | null => {
         return null;
     }
 
-    return (
-        <div css={commonStyles.wrapper}>
-            <div css={commonStyles.contentContainer}>
-                <div css={commonStyles.topLeftComponent}>
-                    <div css={commonStyles.heading}>{header}</div>
-                    <p css={commonStyles.paragraph}>{body}</p>
+    const onCloseClick: OnCloseClick = (evt, internalButtonId) => {
+        evt.preventDefault();
+        onCloseAction(internalButtonId);
+    };
 
-                    <p css={commonStyles.highlightContainer}>
-                        <strong css={commonStyles.highlight}>{cta} &nbsp;</strong>
-                        <span css={styles.storeIcon}>
+    const onCloseAction = (internalButtonId: number): void => {
+        setShowBanner(false);
+        document.body.focus();
+        trackClick({
+            internalButtonId,
+            ophanComponentId: ophanComponentId as string,
+        });
+    };
+
+    return (
+        <div css={styles.wrapper}>
+            <div css={styles.contentContainer}>
+                <div css={styles.topLeftComponent}>
+                    <div css={styles.heading}>{header}</div>
+                    <p css={styles.paragraph}>{body}</p>
+
+                    <p css={styles.highlightContainer}>
+                        <strong css={styles.highlight}>{cta} &nbsp;</strong>
+                        <span css={localStyles.storeIcon}>
                             <AppStore />
                             <PlayStore />
                         </span>
                     </p>
 
                     <ThemeProvider theme={overrridenReaderRevenueTheme}>
-                        <Button
-                            onClick={(e) => onCloseClick(e, 0)}
-                            css={commonStyles.primaryButton}
-                        >
-                            Ok, got it
+                        <Button onClick={(e) => onCloseClick(e, 0)} css={styles.primaryButton}>
+                            {'Ok, got it'}
                         </Button>
                     </ThemeProvider>
                     <ThemeProvider theme={notInterestedTheme}>
@@ -138,28 +115,75 @@ export const AppBanner = (props: Props): ReactElement | null => {
                         </Button>
                     </ThemeProvider>
                 </div>
-                <div css={commonStyles.bottomRightComponent}>
-                    <div css={styles.image}>
+                <div css={styles.bottomRightComponent}>
+                    <div css={localStyles.image}>
                         <img src={imageUrl} alt="" />
                     </div>
-                    <div css={commonStyles.iconPanel}>
-                        <ThemeProvider theme={overrridenReaderRevenueTheme}>
-                            <Button
-                                icon={<SvgCross />}
-                                hideLabel={true}
-                                css={commonStyles.closeButton}
-                                priority="tertiary"
-                                size="small"
-                                aria-label="Close"
-                                onClick={(e) => onCloseClick(e, CLOSE_BUTTON_ID)}
-                                tabIndex={0}
-                            >
-                                {' '}
-                            </Button>
-                        </ThemeProvider>
-                    </div>
+                    <BannerCloseButton
+                        trackClick={trackClick}
+                        setShowBanner={setShowBanner}
+                        ophanComponentId={ophanComponentId}
+                        colors={closeButtonColors}
+                    />
                 </div>
             </div>
         </div>
     );
+};
+
+const defaultColors: BannerColorStyles = {
+    styleBackground: '#ebe8e8',
+    styleHeader: `#333333`,
+    styleBody: '#666',
+    styleHighlight: `#333333`,
+    styleHighlightBackground: '#ebe8e8',
+    styleButton: '#ffffff',
+    styleButtonBackground: '#052962',
+    styleButtonHover: '#234b8a',
+};
+
+const closeButtonColors: CloseButtonColorStyles = {
+    styleClose: `#333333`,
+    styleCloseBackground: '#ebe8e8',
+    styleCloseHover: '#e5e5e5',
+};
+
+const localStyles = {
+    image: css`
+        max-width: 100%;
+        max-height: 260px;
+        display: flex;
+        justify-content: center;
+        align-items: flex-end;
+        margin-top: -20px;
+
+        img {
+            max-width: 100%;
+            width: 100%;
+            height: 100%;
+            object-fit: contain;
+        }
+
+        ${until.desktop} {
+            display: none;
+        }
+
+        ${from.wide} {
+            margin-right: 100px;
+        }
+    `,
+
+    storeIcon: css`
+        height: 10px;
+        display: inline-block;
+        vertical-align: -37.5%;
+
+        svg {
+            height: 30px;
+            width: auto;
+            max-width: 110px;
+            margin-top: ${space[2]}px;
+            padding-right: ${space[2]}px;
+        }
+    `,
 };
