+++ src/styles/colorData.ts
@@ -33,15 +33,18 @@ export interface ReminderButtonColorStyles extends LoadingDotsColorStyles {
     styleReminderButtonHover: ColorValueHex;
 }
 
+export interface CloseButtonColorStyles {
+    styleClose: ColorValueHex;
+    styleCloseBackground: ColorValueHex;
+    styleCloseHover: ColorValueHex;
+}
+
 export interface BannerCopyColorStyles {
     styleBackground: ColorValueHex;
     styleHeader: ColorValueHex;
     styleBody: ColorValueHex;
     styleHighlight: ColorValueHex;
     styleHighlightBackground: ColorValueHex;
-    styleClose: ColorValueHex;
-    styleCloseBackground: ColorValueHex;
-    styleCloseHover: ColorValueHex;
 }
 
 export interface BannerColorStyles extends BannerCopyColorStyles, PrimaryButtonColorStyles {}
@@ -55,11 +58,12 @@ export interface StyleableBannerColorStyles
 export type EpicColorStyles = ReminderButtonColorStyles;
 
 interface AllAvailableColorStyles
-    extends LoadingDotsColorStyles,
-        ReminderButtonColorStyles,
-        PrimaryButtonColorStyles,
+    extends EpicColorStyles,
         BannerColorStyles,
-        EpicColorStyles {}
+        LoadingDotsColorStyles,
+        PrimaryButtonColorStyles,
+        ReminderButtonColorStyles,
+        CloseButtonColorStyles {}
 
 type ColorStylesType = keyof AllAvailableColorStyles;
